feat: home page and fixed search
