+++ src/containers/Cart/index.css
@@ -1,0 +1,20 @@
+div.item-total-details {
+  width: 100%;
+  box-shadow: rgba(0, 0, 0, 0.05) 0px 6px 24px 0px,
+    rgba(0, 0, 0, 0.08) 0px 0px 0px 1px;
+}
+
+div.grid-2.main-section.cart {
+  grid-template-columns: 2fr 4fr;
+  justify-content: flex-start;
+  align-content: flex-start;
+}
+
+button.cart-pay-btn {
+  justify-self: center;
+}
+
+div.flex.column.item-container {
+  align-self: flex-start;
+  justify-self: flex-start;
+}
