+++ src/containers/Home/index.js
@@ -1,225 +1,36 @@
-import { useEffect, useRef, useState } from "react";
+import React from "react";
 import { useNavigate } from "react-router-dom";
-import { useProducts } from "./productContext";
-import { useHeader } from "../Header/headerContext";
-
+import { PRODUCTS_LINK } from "../../utils/constants";
 import "./index.css";
-import {
-  addToWishlist,
-  getAllWishlistProducts,
-} from "../../utils/databaseQuery";
-import { supabaseAuthId } from "../../utils/supabaseClient";
-import { WhishlistProvider } from "../Wishlist/wishlistContext";
 
 const Home = () => {
-  /* Custom webhook for products */
-  const { productFilters, dispatch } = useProducts();
-  const { dispatch: headerDispatch } = useHeader();
-
-  /* States for product filters */
-  const [genderFilter, setGenderFilter] = useState([]);
-  const [starFilter, setStarFilter] = useState([]);
-  const [minPrice, setMinPrice] = useState(1);
-
-  /* Ref for all products */
-  const productsRef = useRef([]);
   const navigate = useNavigate();
-
-  /* Function to handle various filter */
-  const handleFilter = (type) => {
-    if (type === "gender") {
-      return function (e) {
-        if (e.target.checked)
-          setGenderFilter([...genderFilter, e.target.value]);
-        else
-          setGenderFilter([
-            ...genderFilter.filter((item) => item !== e.target.value),
-          ]);
-      };
-    } else if (type === "rating") {
-      return function (e) {
-        if (e.target.checked)
-          setStarFilter([...new Set([...starFilter, +e.target.value[0]])]);
-        else
-          setStarFilter([
-            ...starFilter.filter((item) => item !== +e.target.value[0]),
-          ]);
-      };
-    } else if ("price") {
-      return function (e) {
-        setMinPrice(e.target.value);
-      };
-    }
-  };
-
-  useEffect(() => {
-    dispatch({
-      type: "SET_PRODUCT_FILTERS",
-      payload: { gender: genderFilter, star: starFilter, minPrice },
-    });
-  }, [genderFilter, starFilter, minPrice, dispatch]);
-
+  const featuredProducts = PRODUCTS_LINK;
   return (
-    <div>
-      <main className="main-section grid-2">
-        <aside className="sidebar">
-          <form>
-            <div className="grid-2">
-              <h4 className="h4">Filters</h4>
-              <input
-                type="reset"
-                className="link-btn"
-                onClick={(e) => {
-                  setGenderFilter([]);
-                  setStarFilter([]);
-                  setMinPrice(1);
-                }}
-              ></input>
-            </div>
-
-            <ul className="stacked-list">
-              <li className="text-sm fw-700 no-padding">Category</li>
-              <li className="grid-2 checkbox">
-                <input
-                  type="checkbox"
-                  value="Male"
-                  onClick={handleFilter("gender")}
-                />
-                <label htmlFor="Male">Men</label>
-                <input
-                  type="checkbox"
-                  value="Female"
-                  onClick={handleFilter("gender")}
-                />
-                <label htmlFor="Female">Women</label>
-              </li>
-            </ul>
-
-            <ul className="stacked-list">
-              <li className="text-sm fw-700 no-padding">Ratings</li>
-              <li className="grid-2 checkbox">
-                <input
-                  type="checkbox"
-                  value="4star"
-                  onClick={handleFilter("rating")}
-                />
-                <label htmlFor="4star">4 stars and up</label>
-                <input
-                  type="checkbox"
-                  value="3star"
-                  onClick={handleFilter("rating")}
-                />
-                <label htmlFor="3stars">3 stars and up</label>
-                <input
-                  type="checkbox"
-                  value="2star"
-                  onClick={handleFilter("rating")}
-                />
-                <label htmlFor="2star">2 stars and up</label>
-                <input
-                  type="checkbox"
-                  value="1star"
-                  onClick={handleFilter("rating")}
-                />
-                <label htmlFor="1star">1 star and up</label>
-              </li>
-            </ul>
-            <ul className="price-slider">
-              <li className="text-sm fw-700 no-padding">Price</li>
-              <input
-                type="range"
-                min="1"
-                max="1000"
-                defaultValue={1}
-                className="slider"
-                id="slider"
-                onChange={handleFilter("price")}
+    <div onClick={() => navigate("/products")}>
+      <img
+        className="responsive-img"
+        src="https://assets.myntassets.com/f_webp,w_980,c_limit,fl_progressive,dpr_2.0/assets/images/2022/4/5/de979eef-5aad-43e4-85bf-cdf2f37cb3061649160204949-Sportswear_Desk.jpg"
+        alt="hero-img"
+        srcset=""
+      />
+      <h2 className="h2 gap-header">BIGGEST DEALS ON TOP BRANDS</h2>
+      <div className="grid-auto section home">
+        {featuredProducts.map((item) => {
+          return (
+            <div className="card-wrapper vertical no-padding">
+              <img
+                className="card-img responsive-img"
+                loading="lazy"
+                src={item}
+                alt="Sample Avatar"
               />
-              <span className="price-tip --left">₹ 1</span>
-              <span className="price-tip --right">₹ 1000</span>
-            </ul>
-          </form>
-        </aside>
-        <section className="section grid-auto">
-          {productFilters.map((product, itr) => {
-            return (
-              <div
-                key={product?.id}
-                className="card-wrapper vertical no-padding"
-                ref={(ele) => (productsRef.current[itr] = ele)}
-                onMouseEnter={() => {
-                  const className = productsRef.current[itr].className;
-                  productsRef.current[
-                    itr
-                  ].className = `${className} opacity-low`;
-                  productsRef.current[itr].children[0].style.display = "block";
-                  productsRef.current[itr].children[1].style.display = "block";
-                }}
-                onMouseLeave={() => {
-                  const className = productsRef.current[itr].className;
-                  productsRef.current[itr].className = `${className.replace(
-                    "opacity-low",
-                    ""
-                  )}`;
-                  productsRef.current[itr].children[0].style.display = "none";
-                  productsRef.current[itr].children[1].style.display = "none";
-                }}
-              >
-                <button
-                  id="no-opacity"
-                  className="secondary-btn"
-                  onClick={async () => {
-                    await addToWishlist({
-                      product_id: product.id,
-                      user_id: supabaseAuthId,
-                    });
-                    getAllWishlistProducts().then((wishlist) => {
-                      headerDispatch({
-                        type: "SET_ITEMS",
-                        payload: wishlist.length,
-                      });
-                    });
-                  }}
-                >
-                  Add to Wishlist
-                </button>
-                <button
-                  id="no-opacity"
-                  className="secondary-btn upper"
-                  onClick={() => navigate("/product/" + product?.id)}
-                >
-                  View Details
-                </button>
-                <img
-                  className="card-img responsive-img hero-img"
-                  loading="lazy"
-                  src={product.img}
-                  alt="Sample Avatar"
-                />
-                <section className="flex column flex-start">
-                  <p className="text-md">{product?.brand}</p>
-                  <p className="text-sm sub-text">{product?.brand_sub_text}</p>
-                  <div className="grid-3 pricing">
-                    <h4 className="h4">₹{product?.price}</h4>
-                    <p className="text-sm price-cut sub-text">
-                      ₹{product?.price + product?.price}
-                    </p>
-                    <p className="text-sm discount">(50%)</p>
-                  </div>
-                </section>
-              </div>
-            );
-          })}
-        </section>
-      </main>
+            </div>
+          );
+        })}
+      </div>
     </div>
   );
 };
 
-const HomeProducts = () => (
-  <WhishlistProvider>
-    <Home />
-  </WhishlistProvider>
-);
-
-export default HomeProducts;
+export default Home;
